requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
python-docx>=1.2.0
Pillow>=10.0.0
python-decouple>=3.8
